+++ CONTRIBUTORS.md
@@ -9,6 +9,6 @@ Contributors
 | Jon Perritt | <jon.perritt@rackspace.com>
 | Ash Wilson | <ash.wilson@rackspace.com>
 | Jamie Hannaford | <jamie.hannaford@rackspace.com>
-| Don Schenck | don.schenck@rackspace.com>
-| Gaurav Bajaj | gaurav.bajaj@rackspace.com>
+| Don Schenck | <don.schenck@rackspace.com>
+| Gaurav Bajaj | <gaurav.bajaj@rackspace.com>
 | Joe Topjian | <joe@topjian.net>
